Support recovery of expired jobs in the filesystem queue
This allows for support of either multi-master or forked worker strategies with
the filesystem queue.  When a consumer dies, the jobs will be automatically
recovered based on the timeout.  This is similar to SQS's visibility_timeout
behavior.

This also removes the concept of the standard cleanup API since cleanup should
be occurring periodically as part of the queue's regular operating processes.