Fix the filesystem consumer processing incomplete job files
When the publisher writes jobs to the filesystem, there's a small window of time
where the file is on the filesystem with no data and can be locked by the
consumer.  To avoid reading those files, the consumer will now only process files
that actually have data in them.

With this change, it's now also possible that the publisher can create a file on
the filesystem and then subequently die, leaving a zero-length file that never gets
processed by the consumer.

To avoid leaving empty files on the filesystem forever, the consumer will check for
zero-length files that have exceeded the queue's visibility timeout and automatically
delete them.